File-like interface for working with files in the Workspace (#5)
Add a new module, files.py with a file-like interface emulating the
built-in open for working with files in the Databricks Workspace or in a
Unity Catalog Volume. This should make it easy to write code that uses
these filesystems for persistence which works both locally and in
Databricks.

Clean up some comments in spark.py.